services/transcript/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcript-service"
version = "0.1.0"
description = "Transcript service for Alpha Machine"
authors = [
  { name = "Jonathan", email = "jonathan@example.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "openai",
    "supabase",
    "pyyaml"
]

[project.urls]
"Homepage" = "https://github.com/your-repo/alpha-machine"
"Bug Tracker" = "https://github.com/your-repo/alpha-machine/issues"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"] 